src/main/java/com/evri/interview/service/CourierTransformer.java
package com.evri.interview.service;

import com.evri.interview.model.Courier;
import com.evri.interview.repository.CourierEntity;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.stream.Collectors;

@Component
public class CourierTransformer {

    public Courier toCourier(CourierEntity entity) {
        return Courier.builder()
                .id(entity.getId())
                .name(String.format("%s %s", entity.getFirstName(), entity.getLastName()))
                .active(entity.isActive())
                .build();
    }

    public List<Courier> toCouriers(List<CourierEntity> entities) {
        return entities.stream()
                .map(this::toCourier)
                .collect(Collectors.toList());
    }
}
